Provide a default TimeZone to SqueezeOS based players
This change implements a Slim Query 'getlmstimezone' whose purpose is to
provide SqueezeOS based players with a default TimeZone. The TimeZone is
obtained from 'https://stats.lms-community.org/api/time', and will be based
on the location of the LMS server machine.

A corresponding change has been made to the Community Firmware version of
SqueezeOS.

A player will make the TimeZone request whenever its own TimeZone has not been
initialized. This would be the case on first set up, or following a factory
reset.

This service was formerly provided by MySqueezeBox.com.

There is one small change in behaviour. MSB would have determined its response
from the location of the player, whereas this implementation will determine
its response from the location of the LMS server.